requirements.txt
Django>=3.0,<4.0
djangorestframework
django-extensions
markdown
django-filter
psycopg2>=2.8
numpy
pandas
joblib
sklearn
requests
ipython